pkg/endpoint: check if PolicyMap is nil in syncPolicyMap
If a build fails for an endpoint, the PolicyMap for it will be set to `nil`.
But, the controller that is started to sync the desiredMapState for the endpoint
with the datapath will still run, and thus can result in a panic due to a
nil-dereference of the map. Also add a check if any other fields for endpoint
utilized in this function are nil as well.

Signed-off by: Ian Vernon <ian@cilium.io>